Add exploration endpoints: quality, grid window, variable inspector
- api/exploration.py:
  - GET /datasets/{id}/quality returns the data-quality report.
  - GET /datasets/{id}/grid returns a windowed block of cells (raw or autoscaled),
    with row identifiers; scaled view uses full-column statistics via apply_spec.
  - GET /datasets/{id}/variables/{column} returns the variable inspector (summary,
    histogram, sequence, suggested transform) with a non-destructive transform
    preview via query params.
- Register the exploration router; fix histogram() to return integer counts.
- API tests cover quality, raw/scaled grid windows, and the inspector.

+++ src/scorepilot/core/profiling.py
@@ -118,7 +118,7 @@ def variable_summary(series: pd.Series, column_type: ColumnType | None = None) -
     )
 
 
-def histogram(series: pd.Series, bins: int = 20) -> tuple[list[float], list[float]]:
+def histogram(series: pd.Series, bins: int = 20) -> tuple[list[int], list[float]]:
     """Return ``(counts, edges)`` for a frequency histogram of a numeric column.
 
     Empty or all-missing columns return empty lists.
